Chore(all): Ensure all dependencies are installed
Just a minor enhancement to ensure contributors only need one command to start their local dev ("yarn start" in "docs" or "yarn dev" in "site").

In detail: the "docs" project links directly to the source of "core" and "gallery" so their dependencies must also be installed

This is actually not perfect, since there's a risk that "core" and "gallery" will be using 2 different react instances. It would be better to have a mono workspace I think, to hoist all dependencies