Build: Use distribution in Docs
Previously, the "docs" project import files from "core" and "gallery" directly (i.e. "../../../core"). This let us run only 1 process to watch for changes in all projects. However, this is quite wrong:

- We don't actually use the distribution build, so things may be different between development and distribution.
- The Storybook must be set up to process our source files in core and gallery. This is a big workaround that still work but is very dangerous to leave it there.

After this PR, to work on Moai you'd usually need 2 terminal windows, one for the "docs" and one for the "core". We may use "npm-run-all" to combine them later. However, this is the better way.

As a side note, the "test" project already uses this approach.